requirements.txt
instaloader==4.5.5
instalooter==2.4.4
moviepy==1.0.3
python_dateutil==2.8.1
schedule==0.6.0
google-api-python-client==1.7.2
google-auth==1.8.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
